added directory check before removal and InCopy and InDesign exclusion

+++ Work Scripts (Jamf Pro)/Adobe/Adobe CC/Adobe CC App Upgrade Policy/preinstall.sh
@@ -124,21 +124,38 @@ uninstallResult2017="$?"
 if [[ "$uninstallResult2017" -eq "0" ]]; then
     echo "${appNameForRemoval} 2017 uninstalled"
 fi
-rm -rf "/Applications/${appNameForRemoval} 2017" >/dev/null 2>&1
+sleep 2
+# Confirm the directory has been deleted - manually installed plugins can result in the directory not being removed
+if [[ -d "/Applications/${appNameForRemoval} 2017" ]]; then
+    rm -rf "/Applications/${appNameForRemoval} 2017" >/dev/null 2>&1
+fi
 # Uninstall 2018
 "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2018" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
 uninstallResult2018="$?"
 if [[ "$uninstallResult2018" -eq "0" ]]; then
     echo "${appNameForRemoval} 2018 uninstalled"
 fi
-rm -rf "/Applications/${appNameForRemoval} 2018" >/dev/null 2>&1
-# Uninstall 2019
-"$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2019" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
-uninstallResult2019="$?"
-if [[ "$uninstallResult2019" -eq "0" ]]; then
-    echo "${appNameForRemoval} 2019 uninstalled"
+sleep 2
+# Confirm the directory has been deleted - manually installed plugins can result in the directory not being removed
+if [[ -d "/Applications/${appNameForRemoval} 2018" ]]; then
+    rm -rf "/Applications/${appNameForRemoval} 2018" >/dev/null 2>&1
 fi
-rm -rf "/Applications/${appNameForRemoval} 2019" >/dev/null 2>&1
+# Uninstall 2019
+if [[ "$appNameForRemoval" =~ "InCopy" ]] || [[ "$appNameForRemoval" =~ "InDesign" ]]; then
+    echo "Policy is for ${appNameForRemoval}"
+    echo "If installed, CC 2019 will not be removed"
+else
+    "$binaryPath" --uninstall=1 --sapCode="$sapCode" --baseVersion="$version2019" --platform=osx10-64 --deleteUserPreferences=false >/dev/null 2>&1
+    uninstallResult2019="$?"
+    if [[ "$uninstallResult2019" -eq "0" ]]; then
+        echo "${appNameForRemoval} CC 2019 uninstalled"
+    fi
+    sleep 2
+    # Confirm the directory has been deleted - manually installed plugins can result in the directory not being removed
+    if [[ -d "/Applications/${appNameForRemoval} CC 2019" ]]; then
+        rm -rf "/Applications/${appNameForRemoval} CC 2019" >/dev/null 2>&1
+    fi
+fi
 }
 
 ########################################################################
